+++ test/test_oscar/test_oscar_getter.py
@@ -1,15 +1,15 @@
 from unittest import TestCase
 from datetime import datetime, timezone
 
-from src.pyapnea.oscar.oscar_loader import load_session
-from src.pyapnea.oscar.oscar_getter import get_channel_from_code, event_data_to_dataframe
-from src.pyapnea.oscar.oscar_constants import ChannelID
+from pyapnea.oscar.oscar_loader import load_session
+from pyapnea import get_channel_from_code, event_data_to_dataframe
+from pyapnea import ChannelID
 
 
 class TestOscarSessionGetter(TestCase):
 
     def test_get_channel_from_code(self):
-        filename = '../data/63c6e928.001'
+        filename = '../data/raw/ResMed_1234567890/Events/63c6e928.001'
         oscar_session_data = load_session(filename)
         real_channel = get_channel_from_code(oscar_session_data, ChannelID.CPAP_Te.value)
         expected_channel = oscar_session_data.data.channels[0]
@@ -17,7 +17,7 @@ class TestOscarSessionGetter(TestCase):
         self.assertEqual(expected_channel, real_channel)
 
     def test_data_to_dataframe(self):
-        filename = '../data/63c6e928.001'
+        filename = '../data/raw/ResMed_1234567890/Events/63c6e928.001'
         oscar_session_data = load_session(filename)
         df = event_data_to_dataframe(oscar_session_data, [ChannelID.CPAP_Te.value])
 
@@ -27,7 +27,7 @@ class TestOscarSessionGetter(TestCase):
         self.assertEqual(expected_date, df.loc[0, 'time_utc'].to_pydatetime())
 
     def test_data_to_dataframe_multiple_channels(self):
-        filename = '../data/63c6e928.001'
+        filename = '../data/raw/ResMed_1234567890/Events/63c6e928.001'
         oscar_session_data = load_session(filename)
         df = event_data_to_dataframe(oscar_session_data, [ChannelID.CPAP_Te.value, ChannelID.CPAP_FlowRate.value])
 
@@ -38,19 +38,19 @@ class TestOscarSessionGetter(TestCase):
         self.assertEqual(expected_date, df.loc[0, 'time_utc'].to_pydatetime())
 
     def test_data_to_dataframe_no_channel_found(self):
-        filename = '../data/63c6e928.001'
+        filename = '../data/raw/ResMed_1234567890/Events/63c6e928.001'
         oscar_session_data = load_session(filename)
         df = event_data_to_dataframe(oscar_session_data, [ChannelID.CPAP_AllApnea.value])
         self.assertListEqual(['no_channel'], df.columns.to_list())
 
     def test_data_to_dataframe_empty_channel_list(self):
-        filename = '../data/63c6e928.001'
+        filename = '../data/raw/ResMed_1234567890/Events/63c6e928.001'
         oscar_session_data = load_session(filename)
         df = event_data_to_dataframe(oscar_session_data, [])
         self.assertListEqual(['no_channel'], df.columns.to_list())
 
     def test_data_to_dataframe_multiple_events(self):
-        filename = '../data/61f5f33c.001'
+        filename = '../data/raw/ResMed_1234567890/Events/61f5f33c.001'
         oscar_session_data = load_session(filename)
         flowrate_channel = get_channel_from_code(oscar_session_data, ChannelID.CPAP_FlowRate.value)
 
@@ -65,7 +65,7 @@ class TestOscarSessionGetter(TestCase):
         self.assertFalse(df['FlowRate'].isna().any())
 
     def test_data_to_dataframe_multiple_events_multiple_channels_ignore_nan(self):
-        filename = '../data/61f5f33c.001'
+        filename = '../data/raw/ResMed_1234567890/Events/61f5f33c.001'
         oscar_session_data = load_session(filename)
         flowrate_channel = get_channel_from_code(oscar_session_data, ChannelID.CPAP_FlowRate.value)
 
@@ -83,7 +83,7 @@ class TestOscarSessionGetter(TestCase):
         self.assertFalse(df['FlowRate'].isna().any())
 
     def test_data_to_dataframe_multiple_events_multiple_channels_replace_nan(self):
-        filename = '../data/61f5f33c.001'
+        filename = '../data/raw/ResMed_1234567890/Events/61f5f33c.001'
         oscar_session_data = load_session(filename)
         flowrate_channel = get_channel_from_code(oscar_session_data, ChannelID.CPAP_FlowRate.value)